(dash: core)  isolate getList methods from crud services
